# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API для обработки заявок",
        "title": "Calls service",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/calls": {
            "get": {
                "description": "Retrieves a list of calls belonging to the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calls"
                ],
                "summary": "Get user calls",
                "responses": {
                    "200": {
                        "description": "List of calls",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.CallResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apierrors.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/apierrors.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Saves a new call with client name, phone number, and description",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calls"
                ],
                "summary": "Create a new call",
                "parameters": [
                    {
                        "description": "Call data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.CallDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/apierrors.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apierrors.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/apierrors.Response"
                        }
                    }
                }
            }
        },
        "/calls/{id}": {
            "get": {
                "description": "Retrieves details of a specific call belonging to the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calls"
                ],
                "summary": "Get user call by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Call ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Call details",
                        "schema": {
                            "$ref": "#/definitions/entity.CallResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid call ID",
                        "schema": {
                            "$ref": "#/definitions/apierrors.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apierrors.Response"
                        }
                    },
                    "404": {
                        "description": "Call not found",
                        "schema": {
                            "$ref": "#/definitions/apierrors.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/apierrors.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a call belonging to the authenticated user by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calls"
                ],
                "summary": "Delete call",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Call ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid call ID",
                        "schema": {
                            "$ref": "#/definitions/apierrors.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apierrors.Response"
                        }
                    },
                    "404": {
                        "description": "Call not found or does not belong to user",
                        "schema": {
                            "$ref": "#/definitions/apierrors.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/apierrors.Response"
                        }
                    }
                }
            }
        },
        "/calls/{id}/status": {
            "put": {
                "description": "Updates the status (open or closed) of a specific user call",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calls"
                ],
                "summary": "Update call status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Call ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New status",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.UpdateCallStatusDTO"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid input or status value",
                        "schema": {
                            "$ref": "#/definitions/apierrors.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apierrors.Response"
                        }
                    },
                    "404": {
                        "description": "Call not found or does not belong to user",
                        "schema": {
                            "$ref": "#/definitions/apierrors.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/apierrors.Response"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Authenticates a user and returns a JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "User login credentials",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.AuthRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JWT token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "$ref": "#/definitions/apierrors.Response"
                        }
                    },
                    "401": {
                        "description": "Invalid username or password",
                        "schema": {
                            "$ref": "#/definitions/apierrors.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/apierrors.Response"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "Registers a new user with username and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register user",
                "parameters": [
                    {
                        "description": "User registration data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.AuthRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "$ref": "#/definitions/apierrors.Response"
                        }
                    },
                    "409": {
                        "description": "User already exists",
                        "schema": {
                            "$ref": "#/definitions/apierrors.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/apierrors.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "apierrors.Response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "entity.AuthRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "entity.CallDTO": {
            "type": "object",
            "required": [
                "client_name",
                "description",
                "phone_number"
            ],
            "properties": {
                "client_name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
        "entity.CallResponse": {
            "type": "object",
            "properties": {
                "client_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "phone_number": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "entity.UpdateCallStatusDTO": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        }
    }
}